test-contents/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
awsebcli==3.23.3
blessed==1.21.0
botocore==1.38.11
cement==2.10.14
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==44.0.3
deprecation==2.1.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
frozenlist==1.5.0
gotrue==2.11.4
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jinxed==1.3.0
jmespath==1.0.1
multidict==6.2.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
postgrest==0.19.3
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pythonnet==3.0.5
PyYAML==6.0.2
realtime==2.4.1
requests==2.32.3
requests-oauthlib==2.0.0
semantic-version==2.10.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
